fix(ui): align with registry and environments moves on master
- environments/index.tsx: correct relative paths to one fewer level
  (file at components/environments/, not Lifecycle/Environments/) and
  point Modal/EmptyState to the right locations under shared/.
- layout/Navigator/NavigatorExtension.tsx: registry modals moved to
  components/registry/ in #19314.
- registry/ImportModel.tsx: shared/Modal/Modal is one level up, not two.
- eslint.config.js: refresh legacy allowlist for environments paths
  now under components/environments/.
