+++ ui/components/layout/Navigator/NavigatorExtension.tsx
@@ -27,8 +27,8 @@ import RJSFForm from '../../MesheryMeshInterface/PatternService/RJSF';
 import { DynamicFullScreenLoader } from '../../shared/LoadingState/DynamicFullscreenLoader';
 import Troubleshoot from '../../TroubleshootingComponent';
 import TypingFilter from '../../TypingFilter';
-import CreateModelModal from '../../Settings/Registry/CreateModelModal';
-import ImportModelModal from '../../Settings/Registry/ImportModelModal';
+import CreateModelModal from '../../registry/CreateModelModal';
+import ImportModelModal from '../../registry/ImportModelModal';
 import { ViewInfoModal } from '../../shared/Modal/ViewInfoModal';
 import { selectK8sConfig, selectSelectedK8sClusters } from '@/store/slices/mesheryUi';
 import { useSelector } from 'react-redux';
