+++ ui/eslint.config.js
@@ -41,8 +41,8 @@ const legacyRestrictedImportOffenders = [
   'components/DesignLifeCycle/ValidateDesign.tsx',
   'components/DesignLifeCycle/styles.tsx',
   'components/layout/Header/Header.tsx',
-  'components/Lifecycle/Environments/environment-card.tsx',
-  'components/Lifecycle/Environments/index.tsx',
+  'components/environments/environment-card.tsx',
+  'components/environments/index.tsx',
   'components/workspaces/WorkspaceGridView.tsx',
   'components/MesheryAdapterPlayComponent.tsx',
   'components/MesheryChart.tsx',
@@ -82,9 +82,9 @@ const legacyLiteralColorOffenders = [
   'components/General/error-404/styles.tsx',
   'components/layout/Header/Header.styles.tsx',
   'components/layout/Header/Header.tsx',
-  'components/Lifecycle/Environments/environment-card.tsx',
-  'components/Lifecycle/Environments/index.tsx',
-  'components/Lifecycle/Environments/styles.tsx',
+  'components/environments/environment-card.tsx',
+  'components/environments/index.tsx',
+  'components/environments/styles.tsx',
   'components/Lifecycle/General/empty-state/curvedArrowIcon.tsx',
   'components/Lifecycle/General/empty-state/index.tsx',
   'components/Lifecycle/General/flip-card/index.tsx',
@@ -162,7 +162,7 @@ const legacyMaxLineOffenders = [
 // Allowlisted so CI stays green; entries leave the list as files get split up.
 const legacyMaxLineSoftOffenders = [
   'components/Dashboard/resources/network/config.tsx',
-  'components/Lifecycle/Environments/index.tsx',
+  'components/environments/index.tsx',
   'components/layout/Navigator/Navigator.tsx',
   'components/Performance/PerformanceResults.tsx',
   'components/registry/Stepper/UrlStepper.tsx',
